+++ sctbench/inspect_examples/nedmalloc_test.comb.c
@@ -5,7 +5,7 @@ An example of how to use nedalloc
 
 #include <stdio.h>
 #include <stdlib.h>
-#include <time.h>
+#include <sys/time.h>
 #include <sys/mman.h>